Loading janky animations
Still figuring out skinning

sort of maybe didnt fix skinning

Got animation data loading
- need to do skinning
- need to figure out bind poses

fbx sdk libs

more animation stuff

moving some files around

Animation curves and timespans

FBX SDK to load animations

+++ projects/atLib/source/IO/atScan.cpp
@@ -2,7 +2,7 @@
 // -----------------------------------------------------------------------------
 // The MIT License
 // 
-// Copyright(c) 2018 Michael Batchelor, 
+// Copyright(c) 2020 Michael Batchelor, 
 // 
 // Permission is hereby granted, free of charge, to any person obtaining a copy
 // of this software and associated documentation files(the "Software"), to deal
@@ -138,7 +138,7 @@ static int64_t _ScanInteger(const char *str, int64_t *pLen, const char *candidat
   const int64_t intlen = lastNum - nextChar;
   const int64_t baseChar = candidates[0];
 
-  int64_t mag = (int64_t)pow(nCandidates, intlen - 1);
+  int64_t mag = (int64_t)pow((double)nCandidates, double(intlen - 1));
   int64_t res = 0;
   for (int64_t c = nextChar; c < lastNum; ++c, mag /= nCandidates)
     res += (str[c] - baseChar) * mag;
